% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exData.R
\docType{data}
\name{exData}
\alias{exData}
\title{Example dataset}
\format{
A list with seven objects
\describe{
\item{strain}{a character vector containing the strain names.}
\item{subspecies}{a factor specifying to which subspecie the strains belong to.}
\item{HCN}{a numerical vector with their corresponding median genome wide coverage.}
\item{fastafiles}{a character vector containing the file names of the minicircle sequences in fasta format.}
\item{ucs}{a character vector containing the file names of the cluster information in uc format.}
\item{mapstats}{a character vector containing the files names of the mapping statistics in text format.}
\item{depthstats}{a character vector containing the files names of the depth statistics in text format.}
}
}
\usage{
exData
}
\description{
A dataset containing example inputs.
}
\keyword{datasets}
