% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\docType{data}
\name{matrices}
\alias{matrices}
\title{Example cluster matrices}
\format{
a list of different cluster matrices.
}
\usage{
matrices
}
\description{
A list containing 15 example cluster matrices with percent identities of 80, 85 and 88:100.
}
\keyword{datasets}
