% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.uc.R
\name{read.uc}
\alias{read.uc}
\title{Read in uc files}
\usage{
read.uc(file)
}
\arguments{
\item{file}{the name of the uc file e.g. all.minicircles.circ.id70.uc}
}
\value{
\item{hits}{a table containing all hit records.}
\item{clusters}{a table containing all cluster records.}
\item{clustnumbers}{a vector containing the cluster numbers (0-based).}
}
\description{
Clustering based on a percent identity, performed with VSEARCH, generates files in uc format.
The read.uc function allows you to read in an uc file and store cluster and hit records in individual tables.
}
