\name{rKOMICS}
\alias{rKOMICS}
\docType{package}
\title{
\packageTitle{rKOMICS}
}

\description{
\packageDescription{rKOMICS}
}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{For processing mitochondrial genome assemblies of Kinetoplastid.}
\packageIndices{rKOMICS}
}

\author{
Frederik Van den Broeck \email{frederik.vandenbroeck@kuleuven.be}

Manon Geerts \email{mgeerts@itg.be}
}

\seealso{
Github:   \url{https://frebio.github.io/}

komics-suite:   \url{https://frebio.github.io/komics/}

}

\references{
Van den Broeck F, Savill NJ, Imamura H, Sanders M, Maes I, Cooper S, et al. Ecological divergence and hybridization of Neotropical Leishmania parasites. Proc Natl Acad Sci U S A. 2020;117. \doi{doi:10.1073/pnas.1920136117}.
}